feat: create studio apps from the app on install, delete on uninstall
